moved trade route query
