+++ x/stakeibc/keeper/grpc_query.go
@@ -70,3 +70,14 @@ func (k Keeper) Validators(c context.Context, req *types.QueryGetValidatorsReque
 
 	return &types.QueryGetValidatorsResponse{Validators: hostZone.Validators}, nil
 }
+
+func (k Keeper) AllTradeRoutes(c context.Context, req *types.QueryAllTradeRoutes) (*types.QueryAllTradeRoutesResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "invalid request")
+	}
+	ctx := sdk.UnwrapSDKContext(c)
+
+	routes := k.GetAllTradeRoutes(ctx)
+
+	return &types.QueryAllTradeRoutesResponse{TradeRoutes: routes}, nil
+}
